requirements.txt
riptide-lib >= 0.8.0b1, < 0.9
